Move the form, use BytesIO and remove RadioField
- The form `ExportToExcelWorksheets` is now in a more suitable place
- Simplify code: don't use tempfile.TemporaryFile. Opening a file
with open and mode read-binary already gives a Binary I/O object.
- Superfluous RadioField removed
- clean up comments
